fixed the issue with posts and thier buttons

+++ e-commerce-API/database/migrations/2025_12_03_114955_create_order_items_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
         $table->foreignIdFor(Product::class);
         $table->foreignIdFor(Order::class);
         $table->integer('quantity');
-        $table->float('price_when_purchased');
+        $table->decimal('price_when_purchased', 10, 2);
         $table->timestamps();
       });
     }
